Fix warning for benchmarks

+++ benchmarks/01_trivial/02_soaaos.cpp
@@ -63,7 +63,7 @@ template <typename T> static void check(T *dst, int count) {
 
 #define SOA_TO_AOS_STDLIB(N, T_C, T_ISPC)                                                                              \
     static void soa_to_aos##N##_stdlib_##T_ISPC(benchmark::State &state) {                                             \
-        int count = state.range(0);                                                                                    \
+        int count = static_cast<int>(state.range(0));                                                                  \
         T_C *src = static_cast<T_C *>(aligned_alloc_helper(sizeof(T_C) * count));                                      \
         T_C *dst = static_cast<T_C *>(aligned_alloc_helper(sizeof(T_C) * count));                                      \
         init##N(src, dst, count);                                                                                      \
@@ -80,7 +80,7 @@ template <typename T> static void check(T *dst, int count) {
 
 #define SOA_TO_AOS_ISPC(N, T_C, T_ISPC)                                                                                \
     static void soa_to_aos##N##_ispc_##T_ISPC(benchmark::State &state) {                                               \
-        int count = state.range(0);                                                                                    \
+        int count = static_cast<int>(state.range(0));                                                                  \
         T_C *src = static_cast<T_C *>(aligned_alloc_helper(sizeof(T_C) * count));                                      \
         T_C *dst = static_cast<T_C *>(aligned_alloc_helper(sizeof(T_C) * count));                                      \
         init(src, dst, count);                                                                                         \
